Started using size_t instead of int when possible

+++ sa.h
@@ -10,39 +10,39 @@ using namespace std;
 
 struct RMQ
 {
-    vector<int> a;
-    vector<vector<int> > rmq;
-    RMQ(vector<int> aa)
+    vector<size_t> a;
+    vector<vector<size_t>> rmq;
+    RMQ(vector<size_t> aa)
     {
         a = aa;
         int n = aa.size();
         rmq.resize(n);
-        for (int i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             rmq[i].resize(log(n)+1);
-        for(int i = 0; i<n; i++) rmq[i][0] = i;
+        for(size_t i = 0; i<n; i++) rmq[i][0] = i;
     		for(int j = 1; (1<<j) <= n; j++)
-    			for(int i = 0; i + (1<<j) <= n; i++)
+    			for(size_t i = 0; i + (1<<j) <= n; i++)
     				if(a[rmq[i][j-1]] < a[rmq[i+(1<<(j-1))][j-1]])
     					rmq[i][j] = rmq[i][j-1];
     				else rmq[i][j] = rmq[i+(1<<(j-1))][j-1];
     }
     RMQ(){}
-    int log(int n)
+    size_t log(size_t n)
     {
-        int res = 0;
+        size_t res = 0;
         while(n >= (1<<res)) res++;
         return res-1;
     }
-    int query(int i, int j)
+    size_t query(size_t i, size_t j)
     {
         int k = log(j - i + 1);
     	return min(a[rmq[i][k]], a[rmq[j-(1<<k)+1][k]]);
     }
 };
 
-vector<int> toIntVector(string str) {
-    vector<int> res(str.length() + 3, 0);
-    for (int i = 0; i < str.length(); i++) {
+vector<size_t> toIntVector(string str) {
+    vector<size_t> res(str.length() + 3, 0);
+    for (size_t i = 0; i < str.length(); i++) {
         res[i] = str[i] - 'A' + 1;
     }
     res[str.length()] = res[str.length()+1] = res[str.length()+2] = 0;
@@ -50,43 +50,43 @@ vector<int> toIntVector(string str) {
 }
  
  struct SuffixArray {
-    vector<int> str;
-    vector<int> idx;
-    vector<int> inv;
-    vector<int> lcp;
+    vector<size_t> str;
+    vector<size_t> idx;
+    vector<size_t> inv;
+    vector<size_t> lcp;
     
     RMQ rmq;
 
-    int length;
+    size_t length;
     int letters;
 
-    void radixPass(vector<int> a, vector<int>* b, vector<int> ref, int offset, int n, int letters) {
-        vector<int> cnt(letters+1, 0);
+    void radixPass(vector<size_t> a, vector<size_t>* b, vector<size_t> ref, size_t offset, size_t n, int letters) {
+        vector<size_t> cnt(letters+1, 0);
         for(int i = 0; i<letters+1; i++) cnt[i] = 0;
-        for (int i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             cnt[ref[a[i] + offset]]++;
         }
-        for (int i = 0, sum = 0; i <= letters; i++) {
-            int t = cnt[i];
+        for (size_t i = 0, sum = 0; i <= letters; i++) {
+            size_t t = cnt[i];
             cnt[i] = sum;
             sum += t;
         }
-        for (int i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             (*b)[cnt[ref[a[i] + offset]]++] = a[i];
         }
     }
 
-    void build(vector<int> str, vector<int>* sap, int n, int letters) {
-        int n0 = (n + 2) / 3, n2 = n / 3, n02 = n0 + n2, delta = n0 - (n + 1) / 3;
+    void build(vector<size_t> str, vector<size_t>* sap, size_t n, int letters) {
+        size_t n0 = (n + 2) / 3, n2 = n / 3, n02 = n0 + n2, delta = n0 - (n + 1) / 3;
         
-        vector<int> sa12(n02 + 3, 0);
-        vector<int> s12(n02 + 3, 0);
+        vector<size_t> sa12(n02 + 3, 0);
+        vector<size_t> s12(n02 + 3, 0);
         
-        vector<int> sa0(n0, 0);
-        vector<int> s0(n0, 0);
+        vector<size_t> sa0(n0, 0);
+        vector<size_t> s0(n0, 0);
 
         // Sorting two thirds
-        for (int i = 0, j = 0; i < n + delta; i++) {
+        for (size_t i = 0, j = 0; i < n + delta; i++) {
             if (i % 3 > 0) {
                 s12[j++] = i;
             }
@@ -96,8 +96,8 @@ vector<int> toIntVector(string str) {
         radixPass(s12, &sa12, str, 0, n02, letters);
 
         // Checking if the suffixes are sufficiently sorted
-        int name = 0, c0 = -1, c1 = -1, c2 = -1;
-        for (int i = 0; i < n02; i++) {
+        size_t name = 0, c0 = -1, c1 = -1, c2 = -1;
+        for (size_t i = 0; i < n02; i++) {
             if (str[sa12[i]] != c0 || str[sa12[i] + 1] != c1 || str[sa12[i] + 2] != c2) {
                 name++;
                 c0 = str[sa12[i]];
@@ -114,17 +114,17 @@ vector<int> toIntVector(string str) {
         // Recursively sort if not, generate array if it is
         if (name < n02) {
             build(s12, &sa12, n02, name);
-            for (int i = 0; i < n02; i++) {
+            for (size_t i = 0; i < n02; i++) {
                 s12[sa12[i]] = i + 1;
             }
         } else {
-            for (int i = 0; i < n02; i++) {
+            for (size_t i = 0; i < n02; i++) {
                 sa12[s12[i] - 1] = i;
             }
         }
 
         // Sorting lone third
-        for (int i = 0, j = 0; i < n02; i++) {
+        for (size_t i = 0, j = 0; i < n02; i++) {
             if (sa12[i] < n0) {
                 s0[j++] = 3 * sa12[i];
             }
@@ -132,9 +132,9 @@ vector<int> toIntVector(string str) {
         radixPass(s0, &sa0, str, 0, n0, letters);
 
         // Merge
-        for (int p = 0, t = delta, k = 0; k < n; k++) {
-            int i = sa12[t] < n0 ? sa12[t] * 3 + 1 : (sa12[t] - n0) * 3 + 2;
-            int j = sa0[p];
+        for (size_t p = 0, t = delta, k = 0; k < n; k++) {
+            size_t i = sa12[t] < n0 ? sa12[t] * 3 + 1 : (sa12[t] - n0) * 3 + 2;
+            size_t j = sa0[p];
             if (sa12[t] < n0 ?
                     leq(str[i], s12[sa12[t] + n0], str[j], s12[j / 3]) :
                     leq(str[i], str[i + 1], s12[sa12[t] - n0 + 1],
@@ -163,11 +163,11 @@ vector<int> toIntVector(string str) {
         return (a1 < b1 || (a1 == b1 && leq(a2, a3, b2, b3)));
     }
 
-    vector<int> getLCP() {
-        vector<int> lcp(length - 1, 0);
-        int curr = 0;
-        for (int i = 0; i < length; i++) {
-            int k = inv[i];
+    vector<size_t> getLCP() {
+        vector<size_t> lcp(length - 1, 0);
+        size_t curr = 0;
+        for (size_t i = 0; i < length; i++) {
+            size_t k = inv[i];
             if (k < length - 1) {
                 int j = idx[k + 1];
                 while (i + curr < length && j + curr < length &&
@@ -183,14 +183,14 @@ vector<int> toIntVector(string str) {
         return lcp;
     }
     
-    int queryLcp(int a, int b)
+    size_t queryLcp(size_t a, size_t b)
     {
         if(a == b) return length - a;
-        int x = inv[a], y = inv[b];
+        size_t x = inv[a], y = inv[b];
         return rmq.query(min(x, y), max(x, y)-1);
     }
     
-    SuffixArray(vector<int> st, int ln, int ltrs)
+    SuffixArray(vector<size_t> st, size_t ln, int ltrs)
     {
         length = ln;
         letters = ltrs;
@@ -198,7 +198,7 @@ vector<int> toIntVector(string str) {
         idx.resize(length);
         build(str, &idx, length, letters);
         inv.resize(length);
-        for (int i = 0; i < length; i++) {
+        for (size_t i = 0; i < length; i++) {
             inv[idx[i]] = i;
         }
